API компилятора, исправления ошибок  - Создан класс компилятора для использования в качестве библиотеки  - В LexerStates и Lexer указатели заменены на умные  - Исправлены имена макросов  - Функция reset в семантическом анализаторе  - Мелкие исправления ошибок  - CMakeLists подготовлен для использования в качестве подмодуля

+++ include/pycom/parser/Parser.hpp
@@ -17,7 +17,7 @@ private:
     bool token_matches_any(std::vector<Type> _types);
     void error(std::string _message);
     void next_token();
-    Token forward(int _k);
+    Token forward(size_t _k);
     Token get_token();
     Token check_get_next(Type _type);
